testing/requirements.txt
pytest>=7.0.0
pytest-asyncio>=0.21.0
requests>=2.28.0
httpx>=0.24.0
PyJWT>=2.0.0
psycopg2-binary>=2.9.0
